feat(): various fixes

+++ src/modules/user/usecases/deleteUserById/DeleteUserByIdUsecase.ts
@@ -1,11 +1,26 @@
-import { Usecase } from '@/interfaces/base-usecase';
-import { UserRepository } from '../../repositories/UserRepository';
-import { User } from '../../domain/User';
+import { GenericAppError } from '@/core/logic/AppError';
+import { Result, left, right } from '@/core/logic/Result';
+import { Usecase } from '@/core/domain/Usecase';
 
-export class DeleteUserByIdUsecase implements Usecase<number, User> {
+import { UserRepository } from '../../repositories/UserRepository';
+import { DeleteUserByIdErrors } from './DeleteUserByIdErrors';
+import { DeleteUserByIdResponse } from './DeleteUserByIdResponse';
+
+export class DeleteUserByIdUsecase implements Usecase<string, DeleteUserByIdResponse> {
   constructor(private readonly repository: UserRepository) {}
 
-  async execute(id: number): Promise<User> {
-    return await this.repository.deleteUser(id);
+  async execute(id: string): Promise<DeleteUserByIdResponse> {
+    try {
+      const isExists = await this.repository.findUserById(id);
+
+      if (!isExists) {
+        return left(new DeleteUserByIdErrors.UserNotFoundError(id));
+      }
+
+      await this.repository.deleteUser(id);
+      return right(Result.ok<void>());
+    } catch (error) {
+      return left(new GenericAppError.UnexpectedError(error));
+    }
   }
 }
